Cambio de variables float por int.

+++ CODIGO/Header Files/funciones.h
@@ -14,20 +14,20 @@ FORZADORES
 
 
 //Funciones de estado.
-int estado_espera(float deltaT,float T0);
-int estado_compresor_y_forzadores(float deltaT,float T0);
-int estado_forzadores(float deltaT,float T0);
-int estado_descongelamiento(float deltaT, float T0);
-int estado_error1(float deltaT,float T0);
-int estado_error2(float deltaT, float T0);
+int estado_espera(int deltaT,int T0);
+int estado_compresor_y_forzadores(int deltaT,int T0);
+int estado_forzadores(int deltaT,int T0);
+int estado_descongelamiento(int deltaT, int T0);
+int estado_error1(int deltaT,int T0);
+int estado_error2(int deltaT, int T0);
 
 //Funciones comunicación con MICRO.
-void MICRO_ESPERA();
- void MICRO_COMPYFORZ();
- void MICRO_DEFROST();
- void MICRO_FORZ();
-  void MICRO_DEFAULT();
-  void MICRO_ERROR_2();
-  void MICRO_ERROR_1();
+    void MICRO_ESPERA();
+    void MICRO_COMPYFORZ();
+    void MICRO_DEFROST();
+    void MICRO_FORZ();
+    void MICRO_DEFAULT();
+    void MICRO_ERROR_2();
+    void MICRO_ERROR_1();
